requirements.txt
python-telegram-bot[job-queue]
python-dotenv
pytz
aiosqlite
fpdf2
matplotlib
